Scan Fragment and Attach Images

+++ app/src/main/java/com/example/blackbox/inventory/InventoryDB.java
@@ -42,6 +42,7 @@ public class InventoryDB {
     private int numberOfImages = 0;
     private FirebaseFirestore db;
     private FirebaseStorage storage;
+    private ArrayList<Uri> downloadUris;
 
     /**
      * Initializes the Firestore database and the 'inventory' collection reference.
@@ -84,6 +85,10 @@ public class InventoryDB {
         return inventory;
     }
 
+    public StorageReference getImagesStorageReference(){
+        return imagesStorageReference;
+    }
+
     /**
      * Gets the image collection
      * @return images    a CollectionReference object
@@ -97,6 +102,7 @@ public class InventoryDB {
      *
      * @param item The Item object to be added to the database.
      */
+
     public void addItemToDB(Item item) {
         Map<String, Object> data = generateItemHashMap(item);
         // Generate a custom itemId
@@ -147,7 +153,7 @@ public class InventoryDB {
         String imageName = System.currentTimeMillis() + "_"
                 + lastPathSegment.substring(lastPathSegment.length() - 5);
         StorageReference imageRef = imagesStorageReference.child(imageName);
-
+        downloadUris = new ArrayList<>();
         // Upload the file to Firebase Storage
         imageRef.putFile(imageUri)
                 .addOnSuccessListener((OnSuccessListener) o -> {
@@ -155,6 +161,7 @@ public class InventoryDB {
                     imageRef.getDownloadUrl().addOnSuccessListener(downloadUri -> {
                         // Image uploaded successfully, now add its URL to Firestore
                         addImageUrlToFirestore(downloadUri);
+                        downloadUris.add(downloadUri);
                     });
                 })
                 .addOnFailureListener(e -> {
@@ -337,7 +344,8 @@ public class InventoryDB {
                 .get()
                 .addOnSuccessListener(queryDocumentSnapshots -> {
                     for (QueryDocumentSnapshot document : queryDocumentSnapshots) {
-                        deleteImageDocumentAndStorage(document);
+                        String imageUrl = document.getString("imageUrl");
+                        deleteImageDocumentAndStorage(document, imageUrl);
                     }
                 })
                 .addOnFailureListener(e ->
@@ -349,19 +357,20 @@ public class InventoryDB {
      * Deletes an image document from Firestore and the associated image from Firebase Storage.
      *
      * @param document The QueryDocumentSnapshot representing the image document to be deleted.
+     * @param imageUrl represent the image url to be deleted.
      */
-    private void deleteImageDocumentAndStorage(QueryDocumentSnapshot document) {
-        String imageUrl = document.getString("imageUrl");
+    private void deleteImageDocumentAndStorage(QueryDocumentSnapshot document, String imageUrl) {
         if (imageUrl != null) {
             // Delete the image document from Firestore
-            document.getReference().delete()
-                    .addOnSuccessListener(aVoid ->
-                            Log.d("Firestore", "Image document deleted successfully")
-                    )
-                    .addOnFailureListener(e ->
-                            Log.e("Firestore", "Failed to delete image document: " + e.getMessage())
-                    );
-
+            if(document != null){
+                document.getReference().delete()
+                        .addOnSuccessListener(aVoid ->
+                                Log.d("Firestore", "Image document deleted successfully")
+                        )
+                        .addOnFailureListener(e ->
+                                Log.e("Firestore", "Failed to delete image document: " + e.getMessage())
+                        );
+            }
             // Delete the image from Firebase Storage
             StorageReference storageRef = storage.getReferenceFromUrl(imageUrl);
             storageRef.delete()
@@ -397,6 +406,9 @@ public class InventoryDB {
 
     public int getNumberOfImages(){
         return this.numberOfImages;
+    }
+    public ArrayList<Uri> getDownloadUris(){
+        return this.downloadUris;
     }
 }
